+++ libminifi/include/io/DescriptorStream.h
@@ -66,7 +66,7 @@ class DescriptorStream : public io::BaseStream {
    * @param value value to write
    * @param size size of value
    */
-  int write(const uint8_t *value, int size) override;
+  size_t write(const uint8_t *value, size_t size) override;
 
  private:
   std::recursive_mutex file_lock_;
